+++ CHANGELOG.md
@@ -4,6 +4,14 @@ chuck.js
 Changelog
 ---------
 
+**12/20/11**
+
+1. Updated `Timeline.init()` to remember which dispatch handlers are listening for events.
+
+2. Gave the `Slider` object a `selected` property that turns true on mousedown and false on mouseup.
+
+
+
 **12/19/11**
 
 1. `Character`s now have a `userEvents` property -- an array of events they need to listen for -- and an event handler method named accordingly. So a `Character` that responds to a mouse click needs to define a method named `clickHandler`.